requirements.txt
gino==1.0.1
gino-sanic==0.1.0
marshmallow==3.8.0
PyJWT==1.7.1
pytest==6.1.1
python-decouple==3.3
sanic==19.12.4
SQLAlchemy==1.3.20